session, Util.getSessionUser, Login class for form
email instead of username for validation.

BUG: see below

TODO: email and username are hardcoded because the corresponding fields
in this class are still null after filling out the form -> we need to
fix this ASAP.
* Automatic form generation is needed for the validate method. But the
controller login-action and the validate-action do get NULL as
parameter, although I did everything the same as in the example files of
the VO!

example user (hardcoded):
email: nice@am.com
password: nice
the example user needs to be present in the database for userPopup to
work.

password validation does already work, the only problem is that the
login-field values to not get passed to the controller login method and
the Login.validate() method (they should automatically).