1st part of "styling" the views
Made div. CSS Declarations to let the views look good. Also tested on
mobile phone that it looks good there too.

I came until the filterPopup. All below (according to Mockup) is not yet
done. I will do it Saturday evening.

Please: DON'T TOUCH the CSS Statements and HTML Tags in the Views as
long as it is avoidable.

Content can of course be inserted into the views, but please try to keep
the given HTML Tags & Structure.

2nd part follows Saturday evening.